refactor name_ref_kind.rs
