+++ crates/ra_ide_api/src/lib.rs
@@ -19,7 +19,7 @@ mod feature_flags;
 mod status;
 mod completion;
 mod runnables;
-mod name_ref_kind;
+mod name_kind;
 mod goto_definition;
 mod goto_type_definition;
 mod extend_selection;
