+++ crates/ra_ide_api/src/name_kind.rs
@@ -1,13 +1,13 @@
 //! FIXME: write short doc here
 
-use hir::Either;
+use hir::{Either, FromSource};
+use ra_db::FileId;
 use ra_syntax::{ast, AstNode, AstPtr};
 use test_utils::tested_by;
 
 use crate::db::RootDatabase;
 
 pub enum NameKind {
-    Method(hir::Function),
     Macro(hir::MacroDef),
     FieldAccess(hir::StructField),
     AssocItem(hir::AssocItem),
@@ -29,7 +29,7 @@ pub(crate) fn classify_name_ref(
     if let Some(method_call) = name_ref.syntax().parent().and_then(ast::MethodCallExpr::cast) {
         tested_by!(goto_definition_works_for_methods);
         if let Some(func) = analyzer.resolve_method_call(&method_call) {
-            return Some(Method(func));
+            return Some(AssocItem(func.into()));
         }
     }
 
@@ -59,17 +59,12 @@ pub(crate) fn classify_name_ref(
     if let Some(field_expr) = name_ref.syntax().parent().and_then(ast::RecordField::cast) {
         tested_by!(goto_definition_works_for_record_fields);
 
-        let record_lit = field_expr.syntax().ancestors().find_map(ast::RecordLit::cast);
-
-        if let Some(ty) = record_lit.and_then(|lit| analyzer.type_of(db, &lit.into())) {
-            if let Some((hir::Adt::Struct(s), _)) = ty.as_adt() {
-                let hir_path = hir::Path::from_name_ref(name_ref);
-                let hir_name = hir_path.as_ident().unwrap();
-
-                if let Some(field) = s.field(db, hir_name) {
-                    return Some(FieldAccess(field));
-                }
-            }
+        if let Some(record_lit) = field_expr.syntax().ancestors().find_map(ast::RecordLit::cast) {
+            let variant_def = analyzer.resolve_record_literal(&record_lit)?;
+            let hir_path = hir::Path::from_name_ref(name_ref);
+            let hir_name = hir_path.as_ident()?;
+            let field = variant_def.field(db, hir_name)?;
+            return Some(FieldAccess(field));
         }
     }
 
@@ -93,6 +88,50 @@ pub(crate) fn classify_name_ref(
             };
         }
     }
+
+    None
+}
+
+pub(crate) fn classify_name(
+    db: &RootDatabase,
+    file_id: FileId,
+    name: &ast::Name,
+) -> Option<NameKind> {
+    use NameKind::*;
+
+    let parent = name.syntax().parent()?;
+    let file_id = file_id.into();
+
+    if let Some(pat) = ast::BindPat::cast(parent.clone()) {
+        return Some(Pat(AstPtr::new(&pat)));
+    }
+    if let Some(var) = ast::EnumVariant::cast(parent.clone()) {
+        let src = hir::Source { file_id, ast: var };
+        let var = hir::EnumVariant::from_source(db, src)?;
+        return Some(Def(var.into()));
+    }
+    if let Some(field) = ast::RecordFieldDef::cast(parent.clone()) {
+        let src = hir::Source { file_id, ast: hir::FieldSource::Named(field) };
+        let field = hir::StructField::from_source(db, src)?;
+        return Some(FieldAccess(field));
+    }
+    if let Some(field) = ast::TupleFieldDef::cast(parent.clone()) {
+        let src = hir::Source { file_id, ast: hir::FieldSource::Pos(field) };
+        let field = hir::StructField::from_source(db, src)?;
+        return Some(FieldAccess(field));
+    }
+    if let Some(_) = parent.parent().and_then(ast::ItemList::cast) {
+        let ast = ast::ImplItem::cast(parent.clone())?;
+        let src = hir::Source { file_id, ast };
+        let item = hir::AssocItem::from_source(db, src)?;
+        return Some(AssocItem(item));
+    }
+    if let Some(item) = ast::ModuleItem::cast(parent.clone()) {
+        let src = hir::Source { file_id, ast: item };
+        let def = hir::ModuleDef::from_source(db, src)?;
+        return Some(Def(def));
+    }
+    // FIXME: TYPE_PARAM, ALIAS, MACRO_CALL; Union
 
     None
 }
